requirements.txt
pyyaml==5.3
kubernetes==10.0.1
pyopenssl==19.1.0
